test: use vitest as default in test-utils (refs SFKUI-7787)
