+++ packages/test-utils/jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
     preset: "@forsakringskassan/jest-config-vue",
-    testPathIgnorePatterns: ["/node_modules/", "\\.vitest\\.spec\\.ts$"],
+    testPathIgnorePatterns: ["/node_modules/"],
+    testMatch: ["**/*.jest.spec.ts"],
 };
