+++ packages/test-utils/src/matchers/to-have-focus.spec.ts
@@ -1,127 +1,40 @@
-/* eslint-disable jest/no-conditional-expect -- for testing exceptions */
-
+import { beforeEach, describe, expect, it } from "vitest";
 import { toHaveFocus } from "./to-have-focus";
-
-declare global {
-    /* eslint-disable-next-line @typescript-eslint/no-namespace -- module augmentation */
-    namespace jest {
-        interface Matchers<R> {
-            toHaveFocus(): R;
-        }
-    }
-}
-
-expect.addSnapshotSerializer({
-    test(value: unknown) {
-        return typeof value === "string";
-    },
-    serialize(value: string) {
-        /* eslint-disable-next-line no-control-regex -- expected to replace control regex */
-        return value.replaceAll(/\u001B\[[\d;]*m/g, "");
-    },
-});
 
 expect.extend({
     toHaveFocus,
 });
 
-let element: HTMLElement;
-let detached: HTMLElement;
+describe("toHaveFocus (vitest)", () => {
+    let element: HTMLElement;
+    let detached: HTMLElement;
 
-beforeEach(() => {
-    document.body.innerHTML = "";
-    element = document.createElement("input");
-    detached = document.createElement("input");
-    document.body.append(element);
-});
+    beforeEach(() => {
+        document.body.innerHTML = "";
+        element = document.createElement("input");
+        detached = document.createElement("input");
+        document.body.append(element);
+    });
 
-it("should pass if element is focused", () => {
-    expect.assertions(1);
-    element.focus();
-    expect(element).toHaveFocus();
-});
-
-it("should pass if element is not focused", () => {
-    expect.assertions(1);
-    element.focus();
-    expect(detached).not.toHaveFocus();
-});
-
-it("should fail if element is not focused", () => {
-    expect.assertions(2);
-    try {
+    it("should pass if element is focused", () => {
+        element.focus();
         expect(element).toHaveFocus();
-        /* eslint-disable-next-line @typescript-eslint/no-explicit-any -- assume message will be present */
-    } catch (error: any) {
-        expect(error?.message).toMatchInlineSnapshot(`
-            expect(received).toHaveFocus(expected)
+    });
 
-            Expected element to have focus but another element was focused
+    it("should pass with .not when element is not focused", () => {
+        element.focus();
+        expect(detached).not.toHaveFocus();
+    });
 
-            Expected:
-              "html > body > input"
-            Received:
-              "html > body"
-        `);
-    }
-});
+    it("should produce a failure message when no element is focused", () => {
+        expect(() => expect(detached).toHaveFocus()).toThrow(
+            /another element was focused/,
+        );
+    });
 
-it("should fail if detached element is not focused", () => {
-    expect.assertions(2);
-    try {
-        expect(detached).toHaveFocus();
-        /* eslint-disable-next-line @typescript-eslint/no-explicit-any -- assume message will be present  */
-    } catch (error: any) {
-        expect(error?.message).toMatchInlineSnapshot(`
-            expect(received).toHaveFocus(expected)
-
-            Expected element to have focus but another element was focused
-
-            Expected:
-              "input"
-            Received:
-              "html > body"
-        `);
-    }
-});
-
-it("should fail if element is focused", () => {
-    expect.assertions(2);
-    element.focus();
-    expect(() => {
-        expect(element).not.toHaveFocus();
-    }).toThrowErrorMatchingInlineSnapshot(`Expected element not to have focus`);
-});
-
-it("should handle document.activeElement being null", () => {
-    expect.assertions(2);
-    const root = document.documentElement;
-    document.documentElement.remove();
-    try {
-        expect(element).toHaveFocus();
-        /* eslint-disable-next-line @typescript-eslint/no-explicit-any -- assume message will be present  */
-    } catch (error: any) {
-        expect(error?.message).toMatchInlineSnapshot(`
-            expect(received).toHaveFocus(expected)
-
-            Expected element to have focus but no element was focused
-
-            Expected:
-              "html > body > input"
-            Received:
-              "<null>"
-        `);
-    } finally {
-        document.append(root);
-    }
-});
-
-it("should throw error if expected value is not Element", () => {
-    expect.assertions(1);
-    element.focus();
-    expect(() => {
-        expect("foobar").toHaveFocus();
-    }).toThrowErrorMatchingInlineSnapshot(
-        `Expected value must be Element instance but got "string" instead`,
-    );
+    it("should throw if not an Element", () => {
+        expect(() =>
+            expect("foo" as unknown as Element).toHaveFocus(),
+        ).toThrow();
+    });
 });
